libunwind-aarch64.h: alignas -> __attribute__ aligned
public headers should not require c11 unless necessary; musl does not
define alignas when gcc is set to pre-c11. unw_sigcontext is only
defined if __linux__ anyways, so we can just use __attribute__ aligned.

mempool.c alignas is ok because it is a source file compiled with
libunwind settings, and unwind.h is usually not used with gcc.